consolidate the two NoExtrapPredGens together
